Use ingress instead of port mapping for sidecar test
